refactor: rename '(component)' folders to '_components'
