+++ www/admin/app/dashboard/_components/UserMenu.tsx
@@ -13,7 +13,7 @@ import {
 } from "@mantine/core";
 import { useClickOutside, useDisclosure, useElementSize, useHotkeys } from "@mantine/hooks";
 import { ChevronRightIcon, ArrowLeftOnRectangleIcon } from "@heroicons/react/24/solid";
-import ThemeSwitch from "@admin/app/(components)/ThemeSwitch";
+import ThemeSwitch from "@admin/app/_components/ThemeSwitch";
 import { getBorderColor, getCardColor, getHoverColor } from "@admin/src/utils/colors";
 import type { User } from "api/prisma/types";
