+++ www/admin/app/dashboard/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import { headers } from "next/headers";
-import Menu from "./(components)/Menu";
-import Header from "./(components)/Header";
+import Menu from "./_components/Menu";
+import Header from "./_components/Header";
 import { getServerSession } from "next-auth";
 import { authOptions } from "@admin/src/pages/api/auth/[...nextauth]";
 import { trpc } from "@admin/src/utils/trpc";
