+++ www/admin/app/dashboard/pages/_components/DeletePageModal.tsx
@@ -1,6 +1,6 @@
 import { Alert, Button, Group, Modal, Stack, Text } from "@mantine/core";
 import { ExclamationCircleIcon, TrashIcon } from "@heroicons/react/24/outline";
-import SubmitButton from "@admin/app/(components)/SubmitButton";
+import SubmitButton from "@admin/app/_components/SubmitButton";
 import { trpcReact } from "@admin/src/utils/trpcReact";
 import type { PagesModalProps } from "./PageTree";
