+++ www/admin/app/dashboard/settings/_components/WebsiteSettings.tsx
@@ -8,8 +8,8 @@ import { zodResolver } from "@hookform/resolvers/zod";
 import { z } from "zod";
 import { check } from "language-tags";
 import { trpcReact } from "@admin/src/utils/trpcReact";
-import { Section } from "@admin/app/dashboard/(components)/Section";
-import SubmitButton from "@admin/app/(components)/SubmitButton";
+import { Section } from "@admin/app/dashboard/_components/Section";
+import SubmitButton from "@admin/app/_components/SubmitButton";
 
 const inputSchema = z
 	.object({
