+++ www/admin/app/dashboard/settings/page.tsx
@@ -1,5 +1,5 @@
 import { trpc } from "@admin/src/utils/trpc";
-import WebsiteSettings from "./(components)/WebsiteSettings";
+import WebsiteSettings from "./_components/WebsiteSettings";
 import type { Metadata } from "next";
 
 export const metadata: Metadata = {
