+++ www/api/package.json
@@ -2,7 +2,7 @@
 	"name": "api",
 	"version": "0.0.5",
 	"private": true,
-	"main": "index.ts",
+	"main": "main.ts",
 	"type": "module",
 	"scripts": {
 		"start": "NODE_ENV=production tsx .",
